+++ R/list-files-of-type.R
@@ -9,7 +9,9 @@ list_files_of_type <- function(.path, .file_types) {
   if (.file_types == "detect") {
     .file_type_use <- names(which.max(table(.extensions)))
 
-    cli::cli_alert_info(paste0("Detected type = '", .file_type_use, "'"))
+    stopifnot(.file_type_use %in% c("csv", "sas7bdat", "xpt"))
+
+    cli::cli_alert_info(paste0("Detected file type = '", .file_type_use, "'"))
 
     .extensions_not_file_use <- .extensions[!(.extensions %in% .file_type_use)]
     .extensions_not_file_use_important <- .extensions_not_file_use[.extensions_not_file_use %in% c("csv", "sas7bdat", "xpt")]
@@ -25,8 +27,9 @@ list_files_of_type <- function(.path, .file_types) {
     }
 
   } else {
-    .file_type_use <- gsub(".", "", .file_types, fixed = TRUE)
 
+    .file_type_use <- gsub(".", "", .file_types, fixed = TRUE)
+    cli::cli_alert_info(paste0("User specified file type = '", .file_type_use, "'"))
   }
 
   .out$files <- .files